+++ twoliter/src/tool-crates/testsys/src/lib.rs
@@ -1,3 +1,0 @@
-use include_env_compressed::{include_archive_from_env, Archive};
-
-pub const TESTSYS: Archive = include_archive_from_env!("CARGO_BIN_FILE_TESTSYS");